Add test_benchmark_cluster
